Raspberry pi configuration
In order to make powerline work in the raspberry pi (At least with tmux
and vim) you need to manually install pip by running "apt-get install python pip"
and then install powerline with "pip install powerline-status". For some reason, it
stores the configuration files in a different directory. This configuration points
to that other directory.